New: detach() throws Error on detached nodes
The question was whether the library should silently ignore this error
condition or throw an Error. The decision to throw an Error comes from
the idea that the user of the library must have an assumption on the
state of the DOM when calling detach(). Since that assumption is wrong,
it is safer to throw an Error.